requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# NLP Libraries
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
nltk==3.8.1

# Vector Databases
faiss-cpu==1.7.4
chromadb==0.4.18
sentence-transformers==2.2.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Redis & Queue
redis==5.0.1
celery==5.3.4

# HTTP & Async
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-json-logger==2.0.7

# Multichannel adapters
slack-sdk==3.26.1
pymsteams==0.2.2
twilio==8.11.0

# Security & Rate Limiting
slowapi==0.1.9
python-jose[cryptography]==3.3.0

# Testing & Evaluation
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
